Move logging integration test to its own func
